Refactor into dp_fasttext package and replace dp4py with dependencies
