--- dp4py/README.md
@@ -1,2 +1,0 @@
-# dp4py
-Utility python package for digital publishing. Contains useful logging configs.
