--- dp4py/setup.py
@@ -1,12 +1,0 @@
-from setuptools import setup
-
-setup(
-    name='dp4py',
-    version='0.1.0',
-    packages=['log'],
-    url='',
-    license='',
-    author='sullid',
-    author_email='',
-    description=''
-)
