+++ dp_fasttext/reader.py
@@ -1,7 +1,7 @@
 import abc
 from typing import List
 
-from supervised_models.python.page import Page
+from dp_fasttext.page import Page
 
 
 class DocumentReader(abc.ABC):
