--- log/__init__.py
@@ -1,6 +1,0 @@
-import logging
-import logging.config
-from log.log_config import log_config
-
-
-logging.config.dictConfig(log_config)
