+++ requirements.txt
@@ -11,3 +11,6 @@ typing==3.6.2
 beautifulsoup4==4.6.3
 GitPython==2.1.11
 python_json_logger==0.1.9
+git+https://github.com/ONSdigital/dp4py-config.git@master#egg=dp4py_config
+git+https://github.com/ONSdigital/dp4py-logging.git@master#egg=dp4py_logging
+git+https://github.com/ONSdigital/dp4py-sanic.git@master#egg=dp4py_sanic
